refactor: extract postFairness to API.js
